Fixup linter errors
