+++ include/motor-control/core/utils.hpp
@@ -7,20 +7,10 @@
  * functions to account for different radix positions.
  */
 
-sq0_31 convert_to_fixed_point(float value, int to_radix) {
-    return static_cast<int32_t>(value * static_cast<float>(1LL << to_radix));
-}
+auto convert_to_fixed_point(float value, int to_radix) -> sq0_31;
 
-sq31_31 convert_to_fixed_point_64_bit(float value, int to_radix) {
-    return static_cast<int64_t>(value * static_cast<float>(1LL << to_radix));
-}
+auto convert_to_fixed_point_64_bit(float value, int to_radix) -> sq31_31;
 
-sq0_31 fixed_point_multiply(sq0_31 a, sq0_31 b) {
-    int64_t resultant = static_cast<int64_t>(a) * static_cast<int64_t>(b);
-    return static_cast<sq0_31>((resultant >> 31) & 0xFFFFFFFF);
-}
+auto fixed_point_multiply(sq0_31 a, sq0_31 b) -> sq0_31;
 
-sq0_31 fixed_point_multiply(sq31_31 a, sq0_31 b) {
-    int64_t resultant = a * static_cast<int64_t>(b);
-    return static_cast<sq0_31>((resultant >> 31) & 0xFFFFFFFF);
-}
+auto fixed_point_multiply(sq31_31 a, sq0_31 b) -> sq0_31;
